🚚 move: 카테고리 컴포넌트 분리 (#8)
